@@ -1,38 +1,37 @@
 {
-    "name" : "bretrzaun/statuspage",
-    "description": "Adds a status page to an app",
-    "license": "MIT",
-    "authors" : [
-        {
-            "name" : "Bret R. Zaun",
-            "email" : "bretrzaun@hotmail.com"
-        }
-    ],
-    "autoload" : {
-        "psr-4" : {
-            "BretRZaun\\StatusPage\\" : "src/"
-        }
-    },
-    "require": {
-        "php" : ">=7.1",
-        "twig/twig": "^1.23|^2.0",
-        "guzzlehttp/guzzle": "^6.0"
-    },
-    "require-dev" : {
-        "silex/silex": "^2.0",
-        "phpunit/phpunit" : "^7.0",
-        "symfony/browser-kit": "^3.3",
-        "symfony/css-selector": "^3.3",
-        "doctrine/dbal": "~2.2",
-        "elasticsearch/elasticsearch": "*"
-    },
-    "suggest": {
-        "doctrine/dbal" : "for checking DB connections",
-        "elasticsearch/elasticsearch" : "for checking Elasticsearch"
-    },
-    "config" : {
-        "platform": {
-            "php" : "7.1"
-        }
+  "name": "bretrzaun/statuspage",
+  "description": "Adds a status page to an app",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Bret R. Zaun",
+      "email": "bretrzaun@hotmail.com"
     }
+  ],
+  "autoload": {
+    "psr-4": {
+      "BretRZaun\\StatusPage\\": "src/"
+    }
+  },
+  "require": {
+    "php": ">=7.2.5",
+    "twig/twig": "^2.0|^3.0",
+    "guzzlehttp/guzzle": "^6.0"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^8.0",
+    "symfony/browser-kit": "^5.0",
+    "symfony/css-selector": "^5.0",
+    "doctrine/dbal": "~2.2",
+    "elasticsearch/elasticsearch": "*"
+  },
+  "suggest": {
+    "doctrine/dbal": "for checking DB connections",
+    "elasticsearch/elasticsearch": "for checking Elasticsearch"
+  },
+  "config": {
+    "platform": {
+      "php": "7.2.5"
+    }
+  }
 }
